requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==24.2.0
click==8.1.7
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==7.0.1
Markdown==3.5.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0rc1
zipp==3.17.0
